requirements.txt
azure-functions
azure-storage-blob
azure-cognitiveservices-vision-computervision
pyodbc
opencv-python-headless
flask
openai
python-dotenv
gunicorn